Change components structures
