+++ src/components/Editor/parts/Screen/Screen.module.scss
