+++ src/components/Editor/parts/Screen/Screen.tsx
@@ -3,8 +3,8 @@ import { Image, Layer, Stage } from 'react-konva';
 
 import { baseImgSrc, maskImgSrc, streamModeList } from '@/contants/dataset';
 
-import styles from '@/components/Screen/Screen.module.scss';
-import { createImage } from '@/components/Screen/create-image';
+import styles from '@/components/Editor/parts/Screen/Screen.module.scss';
+import { createImage } from '@/components/Editor/parts/Screen/create-image';
 import {
   BlindfoldLayer,
   CommentLayer,
@@ -14,7 +14,7 @@ import {
   RibbonLayer,
   TelopLayer,
   TimeLayer,
-} from '@/components/Screen/parts';
+} from '@/components/Editor/parts/Screen/parts';
 
 import cursorImg from '@/images/cursor.png';
