+++ src/components/Editor/parts/Screen/parts/BlindfoldLayer.tsx
