+++ src/components/Editor/parts/index.ts
@@ -1,0 +1,5 @@
+export { InputComment } from './form-controls/InputComment';
+export { InputPositionAdjuster } from './form-controls/InputPositionAdjuster';
+export { InputTime } from './form-controls/InputTime';
+export { Select, Textarea, Input } from '@/components/Editor/parts/form-controls/parts';
+export * from './Screen/Screen';
