--- src/index.module.scss
@@ -1,4 +1,0 @@
-.content {
-  min-height: 100vh;
-  padding: 100px 20px;
-}
